Fix import statement (I couldn't help it, see desc)
Somehow the import statement in question got reverted to how it was initially.
I thought it might be because of my formatter but I'm not sure now. I've turned
off autoformatting so no weird things should happen anymore. The formatter
helped with quickly fixing previous linting errors but won't be needed in the
future anyways I expect.